+++ packages/tracing/src/browser/web-vitals/getFID.ts
