+++ packages/tracing/src/browser/web-vitals/getLCP.ts
